requirements.txt
streamlit == 1.35.0
pandas == 2.2.2
numpy == 1.26.4
matplotlib == 3.8.4
seaborn == 0.13.2
Pillow == 10.3.0
